WIP: Bouncer class
